requirements.txt
oci>=2.100.0
boto3>=1.26.0
PyYAML>=6.0
